python/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
plaid_python==24.0.0
python-dotenv==0.15.0
